+++ changelog.md
@@ -1,9 +1,11 @@
-# dspic33ck-curiosity-sent-loopback-demo v1.00
+# dspic33ck-curiosity-sent-loopback-demo v1.0.0
 ### Release Highlights
 
-
+ - Initial release : Refer README.md for summary and execution procedure and project added for SENT with data transmission and reception on same board with loopback mechanism.
 
 ### Features Added\Updated
 
+ - SENT communication using dsPIC33CK256MP508.
+ - Asynchronous and Synchronous mode of transmission and reception .